requirements.txt
Flask==2.0.3
Werkzeug==2.0.3  # Explicitly pin Werkzeug version
Flask-Session==0.4.0
mysql-connector-python==8.0.26
redis==4.1.0
python-dotenv==0.19.0
